Allow to index parsed content in the file field
In the README it is noted that the tika result is indexed in the `file`
field. As I had a use case where I need to analyze the tika result in
different ways I was trying to do exactly that but it did not work as
advertised. Although at first I was able to get around it I now need to
be able to do that. The result is this pull request. Part of me hopes,
that I just missed something, another part hopes that the work was not
redundant.

When reading the code I noticed that the tika result is indexed directly
in the base field name (`AttachmentMapper.contentMapper`) and not in the
`file` field. So I went on and added this feature. Obviously it can be
expensive to index the tika result multiple times (we have PDFs of
several MB) so I wanted this to be optional.

My first attempt was to add another index setting but this would have
meant that all fields of the `attachment` type would be indexed like
this. The version now is basically checking whether there exists a
`fields.file` mapping and only then indexes the content again in
it. This is not really explicit so I am basically hoping for some
feedback on how to make it explicit or even if it is ok the way it is
implemented.

For testing purposes I have copied the
`MultipleAttachmentIntegrationTest` to a
`MultipleAttachmentWithFileIntegrationTest` that is using the
`test-mapping-file.json` mapping. The index contains a `plain` analyzer
only with a `whitespace` tokenizer and the mapping for the `file`
sub-field uses that to analyze the content:

```javascript
{
    "my_attachment": {
        "type": "attachment",
        "fields": {
            "file": {"type": "string", "analyzer": "plain"}
        }
    }
}
```

This pull request is for the 1.3 branch which is the most important one
for me regarding this pull request for now. I am very happy to provide
another one for master if you like the idea and way of implementation of
this one.

Unfortunately I will have to pull a Leeroy Jenkins on this one and
probably bring this to a live environment this week...